+++ templates/app-ts/routes/example/index.ts
