+++ templates/app-ts/test/routes/example.test.ts
@@ -7,5 +7,6 @@ test('example is loaded', async (t) => {
   const res = await app.inject({
     url: '/example'
   })
+
   t.equal(res.payload, 'this is an example')
 })
