+++ templates/app-ts/tsconfig.json
@@ -15,7 +15,7 @@
         }
     },
     "include": [
-        "services/**/*.ts",
+        "routes/**/*.ts",
         "plugins/**/*.ts",
         "app.ts"
     ]
