+++ templates/app/routes/example/index.js
@@ -1,0 +1,7 @@
+'use strict'
+
+module.exports = async function (fastify, opts) {
+  fastify.get('/', async function (request, reply) {
+    return 'this is an example'
+  })
+}
